Tests: Move xapi_edit tests to dedicated file
